pyproject.toml
[tool.poetry]
name = "tgparserbot"
version = "0.1.0"
description = "Telegram bot that parses jokes from websites"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.0.0"
aiohttp = "^3.9.0"
beautifulsoup4 = "^4.12.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tg-jokes-bot = "app.main:cli"
